KernelCTF submission verification for android

+++ kernelctf/repro/repro.sh
@@ -34,34 +34,40 @@ fi
 
 echo $CMDLINE
 
-expect -c '
-    set timeout -1
-    set stty_init raw
+if [[ $RELEASE_ID =~ android-[0-9]{2}-x64-[0-9]{8} ]]; then
+    expect -c "../cuttlefish.sh --release_path=$RELEASE_ID --bin_path=exploit_$RELEASE_ID --flag_fn=flag" | tee $QEMU_TXT | sed $'s/\r//' &
+    QEMU_PID="$!"
+elif [[ $RELEASE_ID =~ ^((lts-[0-9]+.[0-9]+(\\.[0-9]+)?)|(mitigation-(v[0-9]+[a-z]?-)?[0-9]+\\.[0-9]+(\\.[0-9]+)?)|(cos-[0-9]{2,}-[0-9]+\\.[0-9]+\\.[0-9]+)|extra-.* ]]; then
+    expect -c '
+        set timeout -1
+        set stty_init raw
 
-    spawn qemu-system-x86_64 -m 3.5G -nographic \
-    -monitor none \
-    -enable-kvm -cpu host,-avx512f -smp cores=2 \
-    -kernel bzImage \
-    -nic user,model=virtio-net-pci \
-    -drive file=rootfs.img,if=virtio,cache=none,aio=native,format=raw,discard=on,readonly=on \
-    -drive file=flag,if=virtio,format=raw,readonly=on \
-    -virtfs local,path=init,mount_tag=init,security_model=none,readonly=on \
-    -virtfs local,path=exp,mount_tag=exp,security_model=none,readonly=on \
-    -append "'"$CMDLINE"'" \
-    -nographic -no-reboot
+        spawn qemu-system-x86_64 -m 3.5G -nographic \
+        -monitor none \
+        -enable-kvm -cpu host,-avx512f -smp cores=2 \
+        -kernel bzImage \
+        -nic user,model=virtio-net-pci \
+        -drive file=rootfs.img,if=virtio,cache=none,aio=native,format=raw,discard=on,readonly=on \
+        -drive file=flag,if=virtio,format=raw,readonly=on \
+        -virtfs local,path=init,mount_tag=init,security_model=none,readonly=on \
+        -virtfs local,path=exp,mount_tag=exp,security_model=none,readonly=on \
+        -append "'"$CMDLINE"'" \
+        -nographic -no-reboot
 
-    expect "# "
-    send "id\n"
+        expect "# "
+        send "id\n"
 
-    expect "# "
-    send "cat /flag\n"
+        expect "# "
+        send "cat /flag\n"
 
-    expect "# "
-    send "exit\n"
+        expect "# "
+        send "exit\n"
 
-    expect eof
-' | tee $QEMU_TXT | sed $'s/\r//' &
-QEMU_PID="$!"
+        expect eof
+    ' | tee $QEMU_TXT | sed $'s/\r//' &
+    QEMU_PID="$!"
+fi
+
 
 while true; do
     # check if qemu.txt modified within $STDOUT_TIMEOUT seconds
